+++ Sources/Domain/Std.swift
@@ -1,4 +1,0 @@
-extension Int: Value {}
-extension Float: Value {}
-extension Double: Value {}
-extension String: Value {}